Rename project into fennec-aurora-task-creator
It is more self-explanatory than "nightlies-watcher"
