+++ tox.ini
@@ -21,7 +21,7 @@ deps =
     asynctest
 
 commands =
-    py.test --cov nightlies_watcher --cov-report term-missing
+    py.test --cov fennec_aurora_task_creator --cov-report term-missing
     coverage html
     flake8
